// src/app/models/vendor-details/raw-business-and-booth-info.model.ts
export interface RawBusinessAndBoothInfo {
  booths_required: number;
  create_time: string;
  direct_marketing_business: boolean;
  food_samples: boolean;
  id: number;
  jewelry_percentage: number;
  not_selling_items: boolean;
  number_extra_chairs: number;
  number_table_and_chair: number;
  premium_lobby: boolean;
  remain_same_booth: boolean;
  tax_permit_number: string;
  update_time: string;
  vendor_comments: string;
  vendor_id: number;
  willing_to_accept_last_minute_booth: boolean;
  willing_to_accept_outside_booth: boolean;
  sponsor_level_id: number;
  inventory_changed_from_prev_year: boolean;
  inventory_changed_comments: string;
}